+++ dataprovider/foxsports/scraper_mlb_batting_box_score.go
@@ -1,4 +1,4 @@
-package eventdata
+package foxsports
 
 import (
 	"encoding/json"
@@ -7,22 +7,51 @@ import (
 	"strings"
 	"time"
 
+	"github.com/lightning-dabbler/sportscrape"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/jsonresponse"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
 	"github.com/lightning-dabbler/sportscrape/util"
 	"github.com/xitongsys/parquet-go/types"
 )
 
-var pitchingHeaders []string = []string{"PITCHERS", "IP", "H", "R", "ER", "BB", "SO", "HR", "ERA"}
+var battingHeaders []string = []string{"BATTERS", "AB", "R", "H", "RBI", "BB", "SO", "LOB", "AVG"}
 
-type MLBPitchingBoxScoreScraper struct {
+// MLBBattingBoxScoreScraperOption defines a configuration option for the scraper
+type MLBBattingBoxScoreScraperOption func(*MLBBattingBoxScoreScraper)
+
+// MLBBattingBoxScoreScraperParams sets the Params option
+func MLBBattingBoxScoreScraperParams(params map[string]string) MLBBattingBoxScoreScraperOption {
+	return func(s *MLBBattingBoxScoreScraper) {
+		s.Params = params
+	}
+}
+
+// NewMLBBattingBoxScoreScraper creates a new MLBBattingBoxScoreScraper with the provided options
+func NewMLBBattingBoxScoreScraper(options ...MLBBattingBoxScoreScraperOption) *MLBBattingBoxScoreScraper {
+	s := &MLBBattingBoxScoreScraper{}
+
+	// Apply all options
+	for _, option := range options {
+		option(s)
+	}
+	s.League = MLB
+	s.Init()
+
+	return s
+}
+
+type MLBBattingBoxScoreScraper struct {
 	EventDataScraper
 }
 
-func (s *MLBPitchingBoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
+func (s MLBBattingBoxScoreScraper) Feed() sportscrape.Feed {
+	return sportscrape.FSMLBBattingBoxScore
+}
+
+func (s *MLBBattingBoxScoreScraper) Scrape(matchup interface{}) sportscrape.EventDataOutput {
 	start := time.Now().UTC()
 	matchupModel := matchup.(model.Matchup)
-	var context Context
+	var context sportscrape.EventDataContext
 	context.AwayTeam = matchupModel.AwayTeamNameFull
 	context.AwayID = matchupModel.AwayTeamID
 	context.HomeTeam = matchupModel.HomeTeamNameFull
@@ -36,7 +65,7 @@ func (s *MLBPitchingBoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
 	url, err := s.ConstructEventDataURL(matchupModel.EventID)
 	if err != nil {
 		log.Println("Issue constructing event data URL")
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	context.URL = url
 	pullTimestamp := time.Now().UTC()
@@ -44,94 +73,94 @@ func (s *MLBPitchingBoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
 	responseBody, err := s.FetchData(url)
 	if err != nil {
 		log.Println("Issue fetching event data")
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	context.PullTimestamp = pullTimestamp
 	// Unmarshal JSON
 	var responsePayload jsonresponse.MLBEventData
 	err = json.Unmarshal(responseBody, &responsePayload)
 	if err != nil {
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	// Check for box score data
 	if responsePayload.BoxScore == nil || responsePayload.BoxScore.BoxScoreSections == nil {
-		log.Printf("No MLB pitching box score data available for event id: %d\n", matchupModel.EventID)
-		return OutputWrapper{Output: data, Context: context}
+		log.Printf("No MLB batting box score data available for event id: %d\n", matchupModel.EventID)
+		return sportscrape.EventDataOutput{Output: data, Context: context}
 	}
 
 	// Check that both Away and Home team box score stats are populated
 	if responsePayload.BoxScore.BoxScoreSections.AwayStats == nil {
-		log.Printf("No MLB pitching box score data available for away team (%s) for event id: %d\n", matchupModel.AwayTeamNameFull, matchupModel.EventID)
-		return OutputWrapper{Output: data, Context: context}
+		log.Printf("No MLB batting box score data available for away team (%s) for event id: %d\n", matchupModel.AwayTeamNameFull, matchupModel.EventID)
+		return sportscrape.EventDataOutput{Output: data, Context: context}
 	}
 
 	if responsePayload.BoxScore.BoxScoreSections.AwayStats == nil {
-		log.Printf("No MLB pitching box score data available for home team (%s) for event id: %d\n", matchupModel.HomeTeamNameFull, matchupModel.EventID)
-		return OutputWrapper{Output: data, Context: context}
+		log.Printf("No MLB batting box score data available for home team (%s) for event id: %d\n", matchupModel.HomeTeamNameFull, matchupModel.EventID)
+		return sportscrape.EventDataOutput{Output: data, Context: context}
 	}
 
-	// validate MLBPitchingBoxScoreStats home and away positions
+	// validate MLBBattingBoxScoreStats home and away positions
 	uriSplit := strings.Split(responsePayload.BoxScore.BoxScoreSections.HomeStats.ContentURI, "/")
 	actualHomeID, err := util.TextToInt64(uriSplit[len(uriSplit)-1])
 	if actualHomeID != matchupModel.HomeTeamID {
 		log.Printf("Home team ID, %d (%s), does not match expected, %d (%s)\n", actualHomeID, responsePayload.BoxScore.BoxScoreSections.HomeStats.Title, matchupModel.HomeTeamID, matchupModel.HomeTeamNameFull)
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 
 	uriSplit = strings.Split(responsePayload.BoxScore.BoxScoreSections.AwayStats.ContentURI, "/")
 	actualAwayID, err := util.TextToInt64(uriSplit[len(uriSplit)-1])
 	if actualAwayID != matchupModel.AwayTeamID {
 		log.Printf("Away team ID, %d (%s), does not match expected, %d (%s)\n", actualAwayID, responsePayload.BoxScore.BoxScoreSections.AwayStats.Title, matchupModel.AwayTeamID, matchupModel.AwayTeamNameFull)
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 
 	// validate headers
-	expectedHeadersSize := len(pitchingHeaders)
+	expectedHeadersSize := len(battingHeaders)
 
-	// validate home pitching headers (index 2)
-	actualHeaders := responsePayload.BoxScore.BoxScoreSections.HomeStats.BoxscoreItems[2].BoxscoreTable.Headers[0].Columns
+	// validate home batting headers (index 0)
+	actualHeaders := responsePayload.BoxScore.BoxScoreSections.HomeStats.BoxscoreItems[0].BoxscoreTable.Headers[0].Columns
 	actualHeaderSize := len(actualHeaders)
 	if actualHeaderSize != expectedHeadersSize {
-		err = fmt.Errorf("Home team pitching headers size mismatch. actual: %d expected: %d", actualHeaderSize, expectedHeadersSize)
-		return OutputWrapper{Error: err, Context: context}
+		err = fmt.Errorf("Home team batting headers size mismatch. actual: %d expected: %d", actualHeaderSize, expectedHeadersSize)
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	for idx, column := range actualHeaders {
-		if column.Text != pitchingHeaders[idx] {
-			err = fmt.Errorf("Home team pitching header '%s' unexpect at index %d. Expected %s", column.Text, idx, pitchingHeaders[idx])
-			return OutputWrapper{Error: err, Context: context}
+		if column.Text != battingHeaders[idx] {
+			err = fmt.Errorf("Home team batting header '%s' unexpect at index %d. Expected %s", column.Text, idx, battingHeaders[idx])
+			return sportscrape.EventDataOutput{Error: err, Context: context}
 		}
 	}
 
-	// validate away pitching headers (index 2)
-	actualHeaders = responsePayload.BoxScore.BoxScoreSections.AwayStats.BoxscoreItems[2].BoxscoreTable.Headers[0].Columns
+	// validate away batting headers (index 0)
+	actualHeaders = responsePayload.BoxScore.BoxScoreSections.AwayStats.BoxscoreItems[0].BoxscoreTable.Headers[0].Columns
 	actualHeaderSize = len(actualHeaders)
 	if actualHeaderSize != expectedHeadersSize {
-		err = fmt.Errorf("Away team pitching headers size mismatch. actual: %d expected: %d", actualHeaderSize, expectedHeadersSize)
-		return OutputWrapper{Error: err, Context: context}
+		err = fmt.Errorf("Away team batting headers size mismatch. actual: %d expected: %d", actualHeaderSize, expectedHeadersSize)
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	for idx, column := range actualHeaders {
-		if column.Text != pitchingHeaders[idx] {
-			err = fmt.Errorf("Away team pitching header '%s' unexpect at index %d. Expected %s", column.Text, idx, pitchingHeaders[idx])
-			return OutputWrapper{Error: err, Context: context}
+		if column.Text != battingHeaders[idx] {
+			err = fmt.Errorf("Away team batting header '%s' unexpect at index %d. Expected %s", column.Text, idx, battingHeaders[idx])
+			return sportscrape.EventDataOutput{Error: err, Context: context}
 		}
 	}
-	stats, err := s.parsePitchingStats(responsePayload, context)
+	stats, err := s.parseBattingStats(responsePayload, context)
 	if err != nil {
-		return OutputWrapper{Error: err, Context: context}
+		return sportscrape.EventDataOutput{Error: err, Context: context}
 	}
 	for _, obj := range stats {
 		data = append(data, *obj)
 	}
 	diff := time.Now().UTC().Sub(start)
 	log.Printf("Scraping of event %d (%s vs %s) completed in %s\n", matchupModel.EventID, matchupModel.AwayTeamNameFull, matchupModel.HomeTeamNameFull, diff)
-	return OutputWrapper{Output: data, Context: context}
+	return sportscrape.EventDataOutput{Output: data, Context: context}
 }
 
-func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresponse.MLBEventData, context Context) ([]*model.MLBPitchingBoxScoreStats, error) {
-	var stats []*model.MLBPitchingBoxScoreStats
+func (s *MLBBattingBoxScoreScraper) parseBattingStats(responsePayload jsonresponse.MLBEventData, context sportscrape.EventDataContext) ([]*model.MLBBattingBoxScoreStats, error) {
+	var stats []*model.MLBBattingBoxScoreStats
 
 	// Home
-	for idx, record := range responsePayload.BoxScore.BoxScoreSections.HomeStats.BoxscoreItems[2].BoxscoreTable.Rows {
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.HomeStats.BoxscoreItems[0].BoxscoreTable.Rows {
 		if record.EntityLink == nil {
 			continue
 		}
@@ -140,7 +169,7 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 			return stats, err
 		}
 
-		statline := &model.MLBPitchingBoxScoreStats{
+		statline := &model.MLBBattingBoxScoreStats{
 			PullTimestamp:        context.PullTimestamp,
 			PullTimestampParquet: types.TimeToTIMESTAMP_MILLIS(context.PullTimestamp, true),
 			EventTime:            context.EventTime,
@@ -151,7 +180,6 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 			Team:                 context.HomeTeam,
 			OpponentID:           context.AwayID,
 			Opponent:             context.AwayTeam,
-			PitchingOrder:        int32(idx + 1),
 		}
 		err = s.parseStatline(statline, record)
 		if err != nil {
@@ -161,7 +189,7 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 	}
 
 	// Away
-	for idx, record := range responsePayload.BoxScore.BoxScoreSections.AwayStats.BoxscoreItems[2].BoxscoreTable.Rows {
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.AwayStats.BoxscoreItems[0].BoxscoreTable.Rows {
 		if record.EntityLink == nil {
 			continue
 		}
@@ -169,7 +197,7 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 		if err != nil {
 			return stats, err
 		}
-		statline := &model.MLBPitchingBoxScoreStats{
+		statline := &model.MLBBattingBoxScoreStats{
 			PullTimestamp:        context.PullTimestamp,
 			PullTimestampParquet: types.TimeToTIMESTAMP_MILLIS(context.PullTimestamp, true),
 			EventTime:            context.EventTime,
@@ -180,7 +208,6 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 			Opponent:             context.HomeTeam,
 			TeamID:               context.AwayID,
 			Team:                 context.AwayTeam,
-			PitchingOrder:        int32(idx + 1),
 		}
 		err = s.parseStatline(statline, record)
 		if err != nil {
@@ -192,43 +219,42 @@ func (s *MLBPitchingBoxScoreScraper) parsePitchingStats(responsePayload jsonresp
 
 }
 
-func (s *MLBPitchingBoxScoreScraper) parseStatline(stats *model.MLBPitchingBoxScoreStats, statline jsonresponse.BoxScoreStatline) error {
+func (s *MLBBattingBoxScoreScraper) parseStatline(stats *model.MLBBattingBoxScoreStats, statline jsonresponse.BoxScoreStatline) error {
 	var err error
 	stats.Player = statline.EntityLink.Player
-	stats.Record = statline.Columns[0].Superscript
-
-	// InningsPitched
+	stats.Position = *statline.Columns[0].Superscript
+	// AtBat
 	if statline.Columns[1].Text == "-" {
-		stats.InningsPitched = 0
+		stats.AtBat = 0
 	} else {
-		stats.InningsPitched, err = util.TextToFloat32(statline.Columns[1].Text)
+		stats.AtBat, err = util.TextToInt32(statline.Columns[1].Text)
 		if err != nil {
 			return err
 		}
 	}
-	// HitsAllowed
+	// Runs
 	if statline.Columns[2].Text == "-" {
-		stats.HitsAllowed = 0
+		stats.Runs = 0
 	} else {
-		stats.HitsAllowed, err = util.TextToInt32(statline.Columns[2].Text)
+		stats.Runs, err = util.TextToInt32(statline.Columns[2].Text)
 		if err != nil {
 			return err
 		}
 	}
-	// RunsAllowed
+	// Hits
 	if statline.Columns[3].Text == "-" {
-		stats.RunsAllowed = 0
+		stats.Hits = 0
 	} else {
-		stats.RunsAllowed, err = util.TextToInt32(statline.Columns[3].Text)
+		stats.Hits, err = util.TextToInt32(statline.Columns[3].Text)
 		if err != nil {
 			return err
 		}
 	}
-	// EarnedRunsAllowed
+	// RunsBattedIn
 	if statline.Columns[4].Text == "-" {
-		stats.EarnedRunsAllowed = 0
+		stats.RunsBattedIn = 0
 	} else {
-		stats.EarnedRunsAllowed, err = util.TextToInt32(statline.Columns[4].Text)
+		stats.RunsBattedIn, err = util.TextToInt32(statline.Columns[4].Text)
 		if err != nil {
 			return err
 		}
@@ -251,20 +277,20 @@ func (s *MLBPitchingBoxScoreScraper) parseStatline(stats *model.MLBPitchingBoxSc
 			return err
 		}
 	}
-	// HomeRunsAllowed
+	// LeftOnBase
 	if statline.Columns[7].Text == "-" {
-		stats.HomeRunsAllowed = 0
+		stats.LeftOnBase = 0
 	} else {
-		stats.HomeRunsAllowed, err = util.TextToInt32(statline.Columns[7].Text)
+		stats.LeftOnBase, err = util.TextToInt32(statline.Columns[7].Text)
 		if err != nil {
 			return err
 		}
 	}
-	// EarnedRunAverage
+	// BattingAverage
 	if statline.Columns[8].Text == "-" {
-		stats.EarnedRunAverage = 0
+		stats.BattingAverage = 0
 	} else {
-		stats.EarnedRunAverage, err = util.TextToFloat32(statline.Columns[8].Text)
+		stats.BattingAverage, err = util.TextToFloat32(statline.Columns[8].Text)
 		if err != nil {
 			return err
 		}