+++ dataprovider/foxsports/scraper_mlb_pitching_box_score_test.go
@@ -1,15 +1,14 @@
 //go:build integration
 
-package eventdata
+package foxsports
 
 import (
 	"log"
 	"path/filepath"
 	"testing"
 
-	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports"
+	"github.com/lightning-dabbler/sportscrape"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
-	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/runner/matchup"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/xitongsys/parquet-go-source/local"
@@ -24,24 +23,32 @@ func TestMLBPitchingBoxScoreScraper(t *testing.T) {
 	}
 
 	// Get matchups
-	matchupRunner := matchup.NewGeneralMatchupRunner(
-		matchup.GeneralMatchupLeague(foxsports.MLB),
-		matchup.GeneralMatchupSegmenter(&foxsports.GeneralSegmenter{Date: "2024-10-30"}),
+	matchupScraper := NewMatchupScraper(
+		MatchupScraperLeague(MLB),
+		MatchupScraperSegmenter(&GeneralSegmenter{Date: "2024-10-30"}),
 	)
 
-	matchups := matchupRunner.GetMatchups()
+	matchuprunner := sportscrape.NewMatchupRunner(
+		sportscrape.MatchupRunnerScraper(matchupScraper),
+	)
+
+	matchups, err := matchuprunner.Run()
+	if err != nil {
+		t.Error(err)
+	}
 
 	// Get boxscore data
-	scraper := MLBPitchingBoxScoreScraper{}
-	scraper.League = foxsports.MLB
-	runner := NewRunner(
-		RunnerName("MLB pitching box score stats"),
-		RunnerConcurrency(1),
-		RunnerScraper(
-			&scraper,
+	boxscoreScraper := NewMLBPitchingBoxScoreScraper()
+	runner := sportscrape.NewEventDataRunner(
+		sportscrape.EventDataRunnerConcurrency(1),
+		sportscrape.EventDataRunnerScraper(
+			boxscoreScraper,
 		),
 	)
-	boxScoreStats := runner.RunEventsDataScraper(matchups...)
+	boxScoreStats, err := runner.Run(matchups...)
+	if err != nil {
+		t.Error(err)
+	}
 	n_stats := len(boxScoreStats)
 	n_expected := 13
 	assert.Equal(t, n_expected, n_stats, "13 statlines")